Move file-walking utilities from lib.rs to files.rs
Closes #130
